requirements.txt
Django>=5.2.8
pandas>=2.0.0
openpyxl>=3.1.0
google-generativeai>=0.3.0
google-genai>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
mongoengine>=0.27.0
pymongo>=4.6.0

